+++ package.json
@@ -20,7 +20,7 @@
     "@emotion/css": "^11.10.5",
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
-    "@team-entry/design_system": "^1.0.42",
+    "@team-entry/design_system": "^1.0.46",
     "@types/react-responsive": "^8.0.5",
     "axios": "^1.2.2",
     "react": "^18.2.0",
