+++ src/components/Main/BoardAtMain.tsx
@@ -2,50 +2,24 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { Text } from '@team-entry/design_system';
 import Arrow from '../../assets/BoardArrow.svg';
-import { useNavigate } from 'react-router-dom';
+import { boardContent } from '../../constant/main';
+import { Link } from 'react-router-dom';
 
 const BoardsAtMain = () => {
-  const navigate = useNavigate();
-
-  const boardContent = [
-    {
-      id: 0,
-      title: '입학 공지사항',
-      onClick: () => navigate('/notice'),
-      content: [
-        '2023학년도 신입생 입학전 과제 제출 안내sdfja;lsdfj',
-        '2023학년도 신입생 오리엔테이션 집행 안내',
-        '2023학년도 신입생 대상 정밀 건강검진 안내',
-        '2023학년도 신입생 2차전형 안내',
-      ],
-    },
-    {
-      id: 1,
-      title: '자주 묻는 질문',
-      onClick: () => navigate('/customer'),
-      content: [
-        '전학 또는 편입으로 입학할 수 있나요?',
-        '자기소개서는 공백을 포함하나요?',
-        '다른 마이스터고에 중복 지원이 가능한가요?',
-        '여런 전형에 중복 지원이 가능한가요?',
-      ],
-    },
-  ];
-
   return (
     <>
       {boardContent.map((item) => (
         <Board key={item.id}>
           <BoardTitle>
-            <Text color="black600" size="title2" onClick={item.onClick} cursor="pointer">
-              {item.title} <img src={Arrow} alt="arrow" />
-            </Text>
+            <Link to={item.link}>
+              <Text color="black600" size="title2">
+                {item.title} <img src={Arrow} alt="arrow" />
+              </Text>
+            </Link>
           </BoardTitle>
           {item.content.map((item, i) => (
-            <BoardContent key={i}>
-              <Text color="realBlack" size="body1">
-                {item.length > 25 ? `${item.slice(0, 24)}...` : item}
-              </Text>
+            <BoardContent key={i} color="realBlack" size="body1">
+              {item}
             </BoardContent>
           ))}
         </Board>
@@ -72,8 +46,12 @@ const BoardTitle = styled.div`
   border-bottom: 1px solid rgba(0, 0, 0, 0.25);
 `;
 
-const BoardContent = styled.div`
+const BoardContent = styled(Text)`
   width: 20.5rem;
   height: 2rem;
   border-bottom: 1px solid rgba(0, 0, 0, 0.1);
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  word-break: break-all;
 `;
