+++ src/constant/main.ts
@@ -1,0 +1,68 @@
+import React from 'react';
+
+export const boardContent = [
+  {
+    id: 0,
+    title: '입학 공지사항',
+    link: '/notice',
+    content: [
+      '2023학년도 신입생 입학전 과제 제출 안내',
+      '2023학년도 신입생 오리엔테이션 집행 안내',
+      '2023학년도 신입생 대상 정밀 건강검진 안내',
+      '2023학년도 신입생 2차전형 안내',
+    ],
+  },
+  {
+    id: 1,
+    title: '자주 묻는 질문',
+    link: 'custumer',
+    content: [
+      '전학 또는 편입으로 입학할 수 있나요?',
+      '자기소개서는 공백을 포함하나요?',
+      '다른 마이스터고에 중복 지원이 가능한가요?',
+      '여런 전형에 중복 지원이 가능한가요?',
+    ],
+  },
+];
+
+export const developers = [
+  { id: 0, name: '이름이름', major: '분야분야' },
+  { id: 1, name: '이름이름', major: '분야분야' },
+  { id: 2, name: '이름이름', major: '분야분야' },
+  { id: 3, name: '이름이름', major: '분야분야' },
+  { id: 4, name: '이름이름', major: '분야분야' },
+  { id: 5, name: '이름이름', major: '분야분야' },
+  { id: 6, name: '이름이름', major: '분야분야' },
+  { id: 7, name: '이름이름', major: '분야분야' },
+  { id: 8, name: '이름이름', major: '분야분야' },
+  { id: 9, name: '이름이름', major: '분야분야' },
+  { id: 10, name: '이름이름', major: '분야분야' },
+  { id: 11, name: '이름이름', major: '분야분야' },
+];
+
+export const shortcut = [
+  {
+    id: 0,
+    link: '/',
+    icon: 'NavigationArrow',
+    title: '입학설명회 참석 예약',
+  },
+  {
+    id: 1,
+    link: '/grade',
+    icon: 'Culculator',
+    title: '성적 산출',
+  },
+  {
+    id: 2,
+    link: '#SchoolVideo',
+    icon: 'Book',
+    title: '학교 소개 ',
+  },
+  {
+    id: 3,
+    link: '#EntryDeveloper',
+    icon: 'ApproveUser',
+    title: 'Entry 개발자 소개',
+  },
+];
